+++ src/KeyVault/KeyVault.Test/PesterTests/ManagedHsmDataPlaneTests.Tests.ps1
@@ -290,14 +290,14 @@ Describe 'Export Import Security Domain by Restored Blob' {
     # Cannot test initializing because it needs a newly created HSM
     It 'Can Download Security Domain ExchangeKey' {
         $exchangeKeyOutputPath = "$PSScriptRoot/ExchangeKey.cer"
-        Import-AzKeyVaultSecurityDomain -Name $hsmName -ExchangeKeyOutputPath $exchangeKeyOutputPath -DownloadExchangeKey 
+        Import-AzKeyVaultSecurityDomain -Name $hsmName -OutFile $exchangeKeyOutputPath -DownloadExchangeKey 
     }
     
     # This command can be executed offline
     It 'Can decrypt and encrypt Security Domain Data locally' {
         $exchangeKeyPath = "$PSScriptRoot/ExchangeKey.cer"
         $SecurityDomainRestoredBlob = "$PSScriptRoot/HsmRestoredBlob.json"
-        Import-AzKeyVaultSecurityDomain -Keys $certsKeys -ExchangeKey  $exchangeKeyPath -SecurityDomainPath $sd.FullName -SecurityDomainRestoredBlob $SecurityDomainRestoredBlob -RestoreBlob 
+        Import-AzKeyVaultSecurityDomain -Keys $certsKeys -ExchangeKeyPath  $exchangeKeyPath -SecurityDomainPath $sd.FullName -OutFile $SecurityDomainRestoredBlob -RestoreBlob 
     }
 
     It 'Can Import Security Domain by Restore Blob' {