+++ src/Accounts/Accounts/ChangeLog.md
@@ -19,6 +19,9 @@
 -->
 
 ## Upcoming Release
+* Fixed an issue that the usage of some cmdlets are not counted.
+
+## Version 2.13.0
 * Supported in-tool notification for version upgrade.
 * Added an alias `Set-AzConfig` to `Update-AzConfig`
 * Refilled credentials from `AzKeyStore` when run `Save-AzContext` [#22355]